Update datetimes.py
This first check that the index has a timezone, then it convert the start_time and end_time to the timezone of the dataframe using the pytz.timezone.localize() method. This way it's making sure that both the dataframe and the start_time, end_time are in the same timezone and then it goes ahead with the existing comparison.